% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer_hit.R
\name{AllOrients}
\alias{AllOrients}
\title{AllOrients}
\usage{
AllOrients(primer)
}
\arguments{
\item{primer}{Character. Primer sequences (e.g., ATGC).}
}
\description{
\code{AllOrients} expands a primer sequence and show all orientations.
}
\examples{
# AllOrients(primer)
}
