% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_obj_handling.R
\name{assign_ps_barcol}
\alias{assign_ps_barcol}
\title{Assign colors to phyloseq::plot_bar object}
\usage{
assign_ps_barcol(
  gg_obj,
  other_name = c("Others", "Undetermined"),
  other_col = c("lightgray", "darkgray"),
  palette = "igv",
  manual_col_code = NULL,
  legend_title = NULL
)
}
\arguments{
\item{gg_obj}{ggplot2 object.}

\item{other_name}{Character. The name of "other" taxa.}

\item{other_col}{Character. The colors to fill "other" taxa.}

\item{palette}{Character. Currently only "igv" is available.}

\item{manual_col_code}{Character vector. The colors to fill "other" taxa.}

\item{legend_title}{Character. The title of the color legend.}
}
\value{
ggplot2 object.
}
\description{
\code{assign_ps_barcol} Assign colors to phyloseq::plot_bar object
}
\examples{
# taxa_name_summarize(ps_obj, "phylum", top_taxa_n = 10)
}
