% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asv_sheet}
\alias{asv_sheet}
\title{Demo ASV table data used for phyloseq object}
\format{
A data frame with 36 rows and 358 variables:
}
\source{
\url{https://doi.org/10.1111/2041-210X.13204}
}
\usage{
asv_sheet
}
\description{
Demo ASV table data used for phyloseq object
}
\keyword{datasets}
