% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_based_rarefaction.R
\name{coverage_info}
\alias{coverage_info}
\title{Obtaining coverage information of a phyloseq object}
\usage{
coverage_info(ps_obj, datatype = "abundance", estimate_div = FALSE)
}
\arguments{
\item{ps_obj}{Phyloseq object.}

\item{datatype}{Character. Specify the data type. \code{abundance} or \code{incidence} (this should be normally \code{abundance} for a phyloseq object).}

\item{estimate_div}{logical. If \code{TRUE}, return the diversity estimate at the minimum coverage using iNEXT::estimateD.}
}
\value{
data.frame or list.
}
\description{
\code{coverage_info} a wrapper function of \code{iNEXT::DataInfo()} for a phyloseq object
}
\details{
\itemize{
\item{Hsieh et al. (2016) Methods in Ecology and Evolution https://doi.org/10.1111/2041-210X.12613}
}
}
\examples{
# coverage_info(ps_obj)
}
