% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_based_rarefaction2.R
\name{coverage_info2}
\alias{coverage_info2}
\title{Calculate coverage information of a phyloseq object}
\usage{
coverage_info2(ps_obj)
}
\arguments{
\item{ps_obj}{Phyloseq object.}
}
\value{
data.frame or list.
}
\description{
\code{coverage_info2} Calculate coverage information of a phyloseq object
}
\examples{
# coverage_info2(ps_obj)
}
