% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_2sp}
\alias{data_2sp}
\title{Two species model time series}
\format{
A data frame with 1000 rows and 3 variables:
\describe{
\item{time}{Time ID}
\item{x}{Species x}
\item{y}{Species y}
}
}
\usage{
data_2sp
}
\description{
Two species model time series
}
\keyword{datasets}
