% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_4sp}
\alias{data_4sp}
\title{Four fish species time series from Ushio et al. (2018)}
\format{
A data frame with 285 rows and 4 variables:
\describe{
\item{Sebastes.inermis}{Japanese red seaperch}
\item{Trachurus.japonicus}{Japanese jack mackerel}
\item{Girella.punctata}{Largescale Blackfish}
\item{Tridentiger.trigonocephalus}{The chameleon goby}
}
}
\source{
\url{https://doi.org/10.1038/nature25504}
}
\usage{
data_4sp
}
\description{
Four fish species time series from Ushio et al. (2018)
}
\keyword{datasets}
