% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_func.R
\name{label_10_to_power}
\alias{label_10_to_power}
\title{Power label function for ggplot2}
\usage{
label_10_to_power(x)
}
\arguments{
\item{x}{Numeric. A numeric variable that will be converted to a scientific notation.}
}
\value{
Scientific notation.
}
\description{
\code{label_10_to_power} transforms a numeric variable to a scientific notation.
}
\examples{
label_10_to_power(100)
}
