% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_wrapper.R
\name{label_to_power10}
\alias{label_to_power10}
\title{Power label function for ggplot2}
\usage{
label_to_power10(x)
}
\arguments{
\item{x}{Numeric. A numeric variable that will be converted to a sceitific notation.}
}
\value{
Scientific notation.
}
\description{
\code{label_to_power10} transforms a numeric variable to a scientific notation.
}
\examples{
label_to_power10(100)
}
