% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macam-package.R
\docType{package}
\name{macam-package}
\alias{macam}
\alias{macam-package}
\title{macam: A collection of miscellaneous R functions for ecology}
\description{
This package contains a collection of miscellaneous functions for ecology. The functions would be useful for ecological studies.
}
\keyword{internal}
