% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_func.R
\name{make_block}
\alias{make_block}
\title{\code{make_block} function from rEDM v0.7.5}
\usage{
make_block(
  block,
  t = NULL,
  max_lag = 3,
  tau = 1,
  lib = NULL,
  restrict_to_lib = TRUE
)
}
\arguments{
\item{block}{a data.frame or matrix where each column is a time series}

\item{t}{Numeric. The time index for the block.}

\item{max_lag}{The total number of lags to include for each variable.}

\item{tau}{The lag to use for time delay embedding.}

\item{lib}{A 2-column matrix (or 2-element vector) where each row specifies the first and last \emph{rows} of the time series to use for attractor reconstruction.}

\item{restrict_to_lib}{Whether to restrict the final lagged block to just the rows specified in lib (if lib exists).}
}
\value{
A data.frame with the lagged columns and a time column. If the original block had columns X, Y, Z and max_lag = 3, then the returned data.frame will have columns TIME, X, X_1, X_2, Y, Y_1, Y_2, Z, Z_1, Z_2.
}
\description{
\code{make_block} generates a lagged block with the appropriate max_lag and tau, while respecting lib (by inserting NANs, when trying to lag past lib regions)
}
\details{
\itemize{
\item{Ye et al. (2018) Applications of Empirical Dynamic Modeling from Time Series. https://github.com/ha0ye/rEDM}
}
}
\keyword{internal}
