% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_folder_handling.R
\name{outdir_create}
\alias{outdir_create}
\title{Create output directory using the currently-running file name}
\usage{
outdir_create(save_dir_name = TRUE, str_end = 3)
}
\arguments{
\item{save_dir_name}{Logical. If TRUE, output directory name is saved.}

\item{str_end}{Integer. Trim \code{str_end} characters from the file name.}
}
\value{
output_dir Output directory name
}
\description{
\code{outdir_create} Create output directory using the currently-running file name
}
\examples{
# outdir_create()
}
