% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_handling.R
\name{plot_rarefy}
\alias{plot_rarefy}
\title{Visualize rarefaction curve}
\usage{
plot_rarefy(ps_obj, plot_rarefied_point = TRUE, ran_seed = 1234)
}
\arguments{
\item{ps_obj}{Phyloseq object.}

\item{plot_rarefied_point}{Logical. Specify whether rarefied reads are plotted.}

\item{ran_seed}{Random seed.}
}
\description{
\code{plot_rarefy} Perform coverage-based rarefaction for a phyloseq object using iNEXT package
}
\examples{
# plot_rarefy(ps_obj)
}
