% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer_hit.R
\name{PrimerHits}
\alias{PrimerHits}
\title{PrimerHits}
\usage{
PrimerHits(primer, fn)
}
\arguments{
\item{primer}{xxx.}

\item{fn}{xxx.}
}
\description{
\code{PrimerHits} xxx
}
\examples{
# PrimerHits(primer, fn)
}
