% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_handling.R
\name{rarefy_even_coverage}
\alias{rarefy_even_coverage}
\title{Perform coverage-based rarefaction for a phyloseq object}
\usage{
rarefy_even_coverage(
  ps_obj,
  coverage = 0.97,
  remove_not_rarefied = FALSE,
  include_iNEXT_results = FALSE,
  nboot = 40,
  knots = 50,
  n_rarefy_iter = 1,
  rarefy_average_method = "round",
  sample_method = "phyloseq",
  ran_seed = 1234
)
}
\arguments{
\item{ps_obj}{Phyloseq object.}

\item{coverage}{User-specified parameter (default = 0.97 = 97\%).}

\item{remove_not_rarefied}{Remove samples of which coverage is lower than coverage.}

\item{include_iNEXT_results}{Include iNEXT results or not. If TRUE, the function returns a list which contains two elements. The first object is a rarefied phyloseq object, and the second object is an iNEXT result. Also, if TRUE, computation time will increase. IfFALSE, it returns a rarefiedphyloseq object only.}

\item{nboot}{Specify nboot of iNEXT function (valid if include_iNEXT_results = TRUE).}

\item{knots}{Specify knots of iNEXT function (valid if include_iNEXT_results = TRUE).}

\item{n_rarefy_iter}{The number of iterations of rarefactions (default = 1).}

\item{rarefy_average_method}{If n_rarefy_iter >= 2, this argument determines how the multiple rarefactions are summarized. rarefy_average_method = "round" uses round(). rarefy_average_method = "floor" uses floor(). rarefy_average_method = "ceiling" uses ceiling().}

\item{sample_method}{Specify which function is used for rarefaction. sample_method = "vegan" uses vegan::rrarefy(), while sample_method = "phyloseq" uses phyloseq:::rarefaction_subsample().}

\item{ran_seed}{Random seed.}
}
\value{
ps_rare Rarefied phyloseq object
}
\description{
\code{rarefy_even_coverage} Perform coverage-based rarefaction for a phyloseq object using iNEXT package
}
\examples{
# rarefy_even_coverage(ps_obj, coverage = 0.97)
}
