% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr_smap.R
\name{s_map_mdr}
\alias{s_map_mdr}
\title{A wrapper function for MDR S-map}
\usage{
s_map_mdr(
  block,
  effect_var,
  E = 0:10,
  tp = -4:4,
  uic_method = "optimal",
  only_uic = FALSE,
  n_ssr = 10000,
  k = floor(sqrt(n_ssr)),
  evaluate_by = "rho",
  max_delay = 2,
  save_smap_coefficients = FALSE,
  random_seed = 1234
)
}
\arguments{
\item{block}{Data.frame contains time series data.}

\item{effect_var}{Character. Column name that indicates effect variable.}

\item{E}{Numeric. Embedding dimensions that will be tested.}

\item{tp}{Numeric. Forecasting time ahead.}

\item{uic_method}{Character. Specify UIC method. \code{optimal} or \code{marginal}. See https://github.com/yutakaos/rUIC for detail.}

\item{only_uic}{Logical. If \code{TRUE}, only UIC results will be returned.}

\item{n_ssr}{Numeric. The total number of embeddings examined.}

\item{k}{Numeric. The number of embeddings used to calculate ensemble distance.}

\item{evaluate_by}{Character. Specify the criteria used to select the top k-th embeddings. Currently only \code{rho} is supported.}

\item{max_delay}{Numeric. The maximum number of delay to be used. This applies only to \code{effect_var}, and this may equal to \code{E-1}. Potential causal variables and their delays are determiend by UIC.}

\item{save_smap_coefficients}{Logical. If \code{TRUE}, S-map coefficients will be saved.}

\item{random_seed}{Numeric. Random seed.}
}
\value{
A list that contains predictions, statistics, and S-map coefficients (if \code{save_smap_coefficients = TRUE})
}
\description{
\code{s_map_mdr} performs MDR S-map introduced in Chang et al. (2021) Ecology Letters.
}
\details{
\itemize{
\item{Chang, C.-W., Miki, T., Ushio, M., Ke, P.-J., Lu, H.-P., Shiah, F.-K. & et al. (2021) Reconstructing large interaction networks from empirical time series data. Ecology Letters, 24, 2763– 2774. https://doi.org/10.1111/ele.13897}
}
}
