% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr_smap.R
\name{s_map_mdr_all}
\alias{s_map_mdr_all}
\title{A all-in-one wrapper function for MDR S-map}
\usage{
s_map_mdr_all(
  block,
  effect_var,
  lib = NULL,
  pred = lib,
  tp = 1,
  E_range = 0:10,
  tp_range = -4:0,
  n_ssr = 100,
  k = 10,
  theta = 0,
  regularized = FALSE,
  lambda = 0,
  alpha = 0,
  glmnet_parallel = FALSE,
  save_smap_coefficients = TRUE,
  silent = FALSE,
  random_seed = 1234
)
}
\arguments{
\item{block}{Data.frame contains time series data. The first column should be the target column.}

\item{effect_var}{Character or Numeric. Column name or index of the effect variable.}

\item{lib}{Numeric vector. Library indices. If \code{NULL}, \code{c(1, nrow(block_mvd))} will be used as \code{lib}.}

\item{pred}{Numeric vector. Prediction indices.}

\item{tp}{Numeric. Forecasting time ahead.}

\item{E_range}{Numeric. Embedding dimensions that will be tested.}

\item{tp_range}{Numeric. \code{tp} tested for UIC.}

\item{n_ssr}{Numeric. The total number of embeddings examined.}

\item{k}{Numeric. The number of embeddings used to calculate ensemble distance.}

\item{theta}{Numeric. Weighing function for S-map.}

\item{regularized}{Logical If \code{TRUE}, regularized S-map will be performed. If \code{FALSE}, the normal S-map will be performed. Please use \code{rEDM::s_map} function.}

\item{lambda}{Numeric. Specify the strength of penalty in the regularization.}

\item{alpha}{Numeric. \code{alpha = 0} is the ridge regression, \code{alpha = 1} is the lasso regression, and \verb{0 < alpha < 1} is an elastic net.}

\item{glmnet_parallel}{Logical. If TRUE, the computation will be parallel (currently, experimental).}

\item{save_smap_coefficients}{Logical. If \code{TRUE}, S-map coefficients will be saved.}

\item{silent}{Logical. if \code{TRUE}, progress message will not be shown.}

\item{random_seed}{Numeric. Random seed.}
}
\value{
A list that contains predictions, statistics, and S-map coefficients (if \code{save_smap_coefficients = TRUE})
}
\description{
\code{s_map_mdr_all} A all-in-one wrapper function for MDR S-map. Detect causality between variables, construct data.frame for multiview embedding, compute multiview distance, and perform MDR S-map. For fine-tuning, use the step-by-step functions such as \code{compute_mvd()} etc..
}
