% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_SSR.R
\name{s_map_regl}
\alias{s_map_regl}
\title{Perform regularized S-map}
\usage{
s_map_regl(
  block_time,
  lib = c(1, NROW(block_time)),
  pred = lib,
  tp = 1,
  target_column = 1,
  lib_column = 1:NCOL(block_time),
  num_neighbors = NCOL(block_time) + 1,
  theta = 0,
  regularized = FALSE,
  lambda = NULL,
  alpha = 0,
  glmnet_parallel = FALSE,
  random_seed = NULL,
  save_smap_coefficients = FALSE
)
}
\arguments{
\item{block_time}{Dataframe or matrix. Original time series.}

\item{lib}{Library indices.}

\item{pred}{Prediction indices.}

\item{tp}{Forecasting time ahead.}

\item{target_column}{Numeric. Indicates target column}

\item{lib_column}{Numeric. Indicates library column}

\item{num_neighbors}{Numeric. The number of nearest neighbors.}

\item{theta}{Numeric. Weighing function for S-map.}

\item{regularized}{Logical If \code{TRUE}, regularized S-map will be performed. If \code{FALSE}, the normal S-map will be performed. Please use \code{rEDM::s_map} function.}

\item{lambda}{Numeric. Specify the strength of penalty in the regularization.}

\item{alpha}{Numeric. \code{alpha = 0} is the ridge regression, \code{alpha = 1} is the lasso regression, and \verb{0 < alpha < 1} is an elastic net.}

\item{glmnet_parallel}{Logical. If TRUE, the computation will be parallel (currently, experimental).}

\item{random_seed}{Numeric. Random seed.}

\item{save_smap_coefficients}{Logical. If \code{TRUE}, S-map coefficients will be saved.}
}
\value{
A list containing:\tabular{ll}{
\code{model_output} \tab  Model predictions \cr
\tab \cr
\code{stats} \tab  Statistics. \cr
\tab \cr
\code{smap_coefficients} \tab  S-map coefficients \cr
}
}
\description{
\code{extended_smap} performs the regularized S-map introduced in Censi et al. (2019) Methods in Ecology and Evolution.
}
\examples{
# s_map_regl()
}
