% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_SSR.R
\name{s_map_rgl}
\alias{s_map_rgl}
\title{Perform regularized S-map}
\usage{
s_map_rgl(
  ts_data,
  E,
  lib = c(1, length(ts_data)),
  pred = lib,
  tp = 1,
  num_neighbors = length(ts_data) + 1,
  theta = 0,
  regularized = FALSE,
  lambda = NULL,
  alpha = 0,
  glmnet_parallel = FALSE,
  random_seed = NULL,
  save_smap_coefficients = FALSE
)
}
\arguments{
\item{ts_data}{Data.frame, matrix, or vector. One-column univariate time series.}

\item{E}{Integer. Embedding dimension.}

\item{lib}{Library indices.}

\item{pred}{Prediction indices.}

\item{tp}{Forecasting time ahead.}

\item{num_neighbors}{Numeric. The number of nearest neighbors.}

\item{theta}{Numeric. Weighing function for S-map.}

\item{regularized}{Logical If \code{TRUE}, regularized S-map will be performed. If \code{FALSE}, the normal S-map will be performed. Please use \code{rEDM::s_map} function.}

\item{lambda}{Numeric. Specify the strength of penalty in the regularization.}

\item{alpha}{Numeric. \code{alpha = 0} is the ridge regression, \code{alpha = 1} is the lasso regression, and \verb{0 < alpha < 1} is an elastic net.}

\item{glmnet_parallel}{Logical. If TRUE, the computation will be parallel (currently, experimental).}

\item{random_seed}{Numeric. Random seed.}

\item{save_smap_coefficients}{Logical. If \code{TRUE}, S-map coefficients will be saved.}
}
\value{
A list containing:\tabular{ll}{
\code{model_output} \tab  Model predictions \cr
\tab \cr
\code{stats} \tab  Statistics. \cr
\tab \cr
\code{smap_coefficients} \tab  S-map coefficients \cr
}
}
\description{
\code{s_map_rgl} is a wrapper function of \code{extended_lnlp()} for regularized S-map. For multivariate S-map, please use \code{extended_lnlp()}.
}
\examples{
# s_map_rgl()
}
