% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_sheet}
\alias{sample_sheet}
\title{Demo sample data used for phyloseq object}
\format{
A data frame with 36 rows and 14 variables:
}
\source{
\url{https://doi.org/10.1111/2041-210X.13204}
}
\usage{
sample_sheet
}
\description{
Demo sample data used for phyloseq object
}
\keyword{datasets}
