% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_dir_handling.R
\name{save_workspace}
\alias{save_workspace}
\title{Save workspace using an output directory name}
\usage{
save_workspace(workspace_dir)
}
\arguments{
\item{workspace_dir}{Character. Workspace will be saved in this directory.}
}
\description{
\code{save_workspace} saves current workspace in a user-specified output directory.
}
\examples{
# output_dir <- outdir_create()
# save_workspace(output_dir)
}
