% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_func.R
\name{singleton_estimator}
\alias{singleton_estimator}
\title{singleton_estimator is a function of obtaining the estimator of singleton based the method of Chiu & Chao (2016)}
\usage{
singleton_estimator(x, datatype = "abundance")
}
\arguments{
\item{x}{a vector of species sample frequencies}

\item{datatype}{Character. Specify the data type. \code{abundance} or \code{incidence} (this should be normally \code{abundance} for a phyloseq object).}
}
\value{
singleton estimate and a numerical vector of corrected data.
}
\description{
singleton_estimator is a function of obtaining the estimator of singleton based the method of Chiu & Chao (2016)
}
\details{
\itemize{
\item{Chiu & Chao (2016) PeerJ https://doi.org/10.7717/peerj.1634}
}
}
