% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tax_sheet}
\alias{tax_sheet}
\title{Demo taxa table data used for phyloseq object}
\format{
A data frame with 358 rows and 13 variables:
}
\source{
\url{https://doi.org/10.1111/2041-210X.13204}
}
\usage{
tax_sheet
}
\description{
Demo taxa table data used for phyloseq object
}
\keyword{datasets}
