% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_obj_handling.R
\name{taxa_name_bundle}
\alias{taxa_name_bundle}
\title{Bundle taxa names in a phyloseq object}
\usage{
taxa_name_bundle(
  ps_obj,
  taxa_rank_bundled,
  top_taxa_n = 10,
  new_taxa_rank = "bundled_tax",
  taxa_rank_list = c("superkingdom", "kingdom", "phylum", "order", "family", "genus",
    "species"),
  add_undet = TRUE
)
}
\arguments{
\item{ps_obj}{Phyloseq object.}

\item{taxa_rank_bundled}{Character. The name of phylogenetic rank that will be bundled.}

\item{top_taxa_n}{Numeric. The number of taxa names that will be retained.}

\item{new_taxa_rank}{Character. The name of the bundled taxa rank. Default is \code{bundled_tax}.}

\item{taxa_rank_list}{Character vector. The specified taxa ranks are used to determine whether a tax is "Undetermined" or not.}

\item{add_undet}{Logical. If TRUE, "Undetermined" is assigned to undetermined taxa. If FALSE, "Others" is assigned.}
}
\value{
Phyloseq object. Bundled taxa names are in the \code{new_taxa_rank} column.
}
\description{
\code{taxa_name_bundle} bundle taxa names in a phylosq object for a visualization purpose. Minor taxa names will be summarized as "Others". Also, if the taxa name is empty, the function assigns "Undetermined" for the taxa.
}
\examples{
# taxa_name_bundle(ps_obj, "phylum", top_taxa_n = 10)
}
