% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin_surrogate.R
\name{twin_surrogate_cpp}
\alias{twin_surrogate_cpp}
\title{Generate twin surrogate time series}
\usage{
twin_surrogate_cpp(
  original,
  dim,
  num.iter,
  tau = 1,
  s = 0.875,
  surrogate.option = c("random", "phase_lock"),
  initial.point = c("same_season", "twins"),
  distance.method = c("norm", "euclid"),
  point.per.year = 24,
  s.update = c("on", "off"),
  n.twin.threshold = 10,
  output.message = F
)
}
\arguments{
\item{original}{Original time series.}

\item{dim}{Integer. Embedding dimension.}

\item{num.iter}{Numeric. The number of surrogate to be generated.}

\item{tau}{Numeric. Time delay in the embedded time series.}

\item{s}{Numeric. Threshold for the twin in the recurrence plot.}

\item{surrogate.option}{Character. If \code{random}, normal twin surrogate will be generated. If \code{phase_lock}, the periodicity of the time series will be preserved.}

\item{initial.point}{Character. If \code{same_season}, twin surrogate time series starts with a point in a same season. If \code{twins}, twin surrogate time series starts with a twin point.}

\item{distance.method}{Character. Specify how the distance between points in a state space is calculated.}

\item{point.per.year}{Numeric. Specify how many observations were made in a unit time (e.g. 24 observations per year).}

\item{s.update}{Character. If \code{TRUE}, update \code{s} if a sufficient number of twins is not found.}

\item{n.twin.threshold}{Numeric. The minimum number of twins to be accepted.}

\item{output.message}{Logical. If \code{TRUE}, show message.}
}
\value{
surrogate.one.col Twin surrogate time series.
}
\description{
\code{twin_surrogate_cpp} generates twin surrogate time series.
}
\examples{
# twin_surrogate_cpp(time_series, 3, 100)
}
