% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr_smap.R
\name{uic_across}
\alias{uic_across}
\title{Perform UIC across data.frame}
\usage{
uic_across(
  block,
  effect_var,
  E_range = 0:10,
  tp_range = -4:0,
  tau = 1,
  num_surr = 1000,
  alpha = 0.05,
  fdr = FALSE,
  silent = FALSE
)
}
\arguments{
\item{block}{Data.frame contains time series data.}

\item{effect_var}{Character or Numeric. Column name or index of the effect variable.}

\item{E_range}{Numeric. Embedding dimensions that will be tested.}

\item{tp_range}{Numeric. \code{tp} tested for UIC.}

\item{tau}{Numeric. \code{tau} tested for UIC.}

\item{num_surr}{Numeric. The number of surrogate data generated to compute p-value.}

\item{alpha}{Numeric. the significant level to determine the embedding dimension of reference model (i.e., \code{E0}). If \code{alpha = NULL}, \code{E0} is set to \code{E - 1}. If \verb{0 < alpha < 1} \code{E0} depends on the model results with lower embedding dimensions. Default is 0.05.}

\item{fdr}{Logical. If \code{TRUE}, calculate False Discovery Rate using \code{stats::p.adjust()} and BH method.}

\item{silent}{Logical. if \code{TRUE}, progress message will not be shown.}
}
\value{
A data.frame that contains UIC results
}
\description{
\code{uic_across} Perform UIC for a target variable and multiple causal variables
}
\details{
\itemize{
\item{Osada & Ushio (2020) rUIC:Unified Information-theoretic Causality for R. https://doi.org/10.5281/zenodo.5163234}
}
}
