% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer_hit.R
\name{AllOrients}
\alias{AllOrients}
\title{AllOrients}
\usage{
AllOrients(primer)
}
\arguments{
\item{primer}{Character. Primer sequences (e.g., ATGC).}
}
\value{
An output contains:\tabular{ll}{
\code{Forward} \tab  Forward sequence \cr
\tab \cr
\code{Complement} \tab  Complement sequence \cr
\tab \cr
\code{Reverse} \tab  Reverse sequence \cr
\tab \cr
\code{RevComp} \tab  Reverse-complement sequence \cr
}
}
\description{
\code{AllOrients} expands a primer sequence and show all orientations.
}
\examples{
# AllOrients(primer)
}
