% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr_smap.R
\name{compute_mvd}
\alias{compute_mvd}
\title{Computing multiview distance}
\usage{
compute_mvd(
  block_mvd,
  effect_var,
  E,
  tp = 1,
  make_block_method = "naive",
  make_block_max_lag = E,
  n_ssr = 10000,
  k = floor(sqrt(n_ssr)),
  random_seed = 1234,
  distance_only = TRUE,
  silent = FALSE
)
}
\arguments{
\item{block_mvd}{Data.frame contains time series data. The first column should be the target column.}

\item{effect_var}{Character or Numeric. Column name or index of the effect variable.}

\item{E}{Numeric. Optimal embedding dimension of \code{effect_var}}

\item{tp}{Numeric. Forecasting time ahead.}

\item{make_block_method}{Character. If \code{naive}, the multivariate data.frame (\code{block_mvd}) is directly used to calculate the multiview distance. If \code{rEDM}, \code{rEDM::make_block()} is used to add the time-delayed ordinate for each variable, which make the method equivalent to Chang et al. (2021).}

\item{make_block_max_lag}{Numeric. \code{max_lag} in \code{rEDM::make_block()}. This argument will be used only if \code{make_block_method = "rEDM"}.}

\item{n_ssr}{Numeric. The total number of embeddings examined.}

\item{k}{Numeric. The number of embeddings used to calculate ensemble distance.}

\item{random_seed}{Numeric. Random seed.}

\item{distance_only}{Logical. if \code{TRUE}, only distance matrix is returned.}

\item{silent}{Logical. if \code{TRUE}, progress message will not be shown.}
}
\value{
A distance matrix and other information (if \code{distance_only = FALSE}).
}
\description{
\code{compute_mvd} Compute multiview distance
}
\details{
\itemize{
\item{Chang, C.-W., Miki, T., Ushio, M., Ke, P.-J., Lu, H.-P., Shiah, F.-K. & et al. (2021) Reconstructing large interaction networks from empirical time series data. Ecology Letters, 24, 2763– 2774. https://doi.org/10.1111/ele.13897}
}
}
