% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr_smap.R
\name{make_block_mvd}
\alias{make_block_mvd}
\title{Generate block for MDR S-map}
\usage{
make_block_mvd(
  block,
  uic_res,
  effect_var,
  max_lag = 0,
  cause_var_colname = "cause_var",
  th_var_colname = "pval",
  p_threshold = 0.05,
  include_var = "strongest_only",
  tp_adjust = NULL,
  sort_tp = TRUE,
  silent = FALSE
)
}
\arguments{
\item{block}{Data.frame contains time series data.}

\item{uic_res}{Data.frame contains UIC results. Usually this is the output of \code{uic_across()}, but custom data.frame may be usable.}

\item{effect_var}{Character or Numeric. Column name or index of the effect variable.}

\item{max_lag}{Numeric. The total number of lags to include for the effect variable. So if max_lag = 3, a variable X will appear with lags \verb{\{X[t], X[t-1], X[t-2]\}}. This argument was previously \code{E_effect_var}, but renamed to clarify its meaning.}

\item{cause_var_colname}{Character. A column name for causal variables in \code{uic_res}.}

\item{th_var_colname}{Character. A column name for a threshold to choose causal variables in \code{uic_res}.}

\item{p_threshold}{Numeric. Rows with pval larger than \code{p_threshold} is removed.}

\item{include_var}{Character. \code{all_significant}, \code{strongest_only}, or \code{tp0_only}. If \code{all_significant}, all significantly influencing variables are used for the embedding. If \code{strongest_only}, tp with the strongest influence for each variable is used. If \code{tp0_only}, only variables with no time-delay are used.}

\item{tp_adjust}{NULL or numeric. Adjust tp values.}

\item{sort_tp}{Logical. If TRUE, \code{block_mvd} is sorted according to \code{tp} for each causal variable.}

\item{silent}{Logical. if \code{TRUE}, progress message will not be shown.}
}
\value{
Embedded time series will be returned.
}
\description{
\code{make_block_mvd} Generate data.frame for MDR S-map
}
