% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr_smap.R
\name{make_block_mvd}
\alias{make_block_mvd}
\title{Generate block for MDR S-map}
\usage{
make_block_mvd(
  block,
  uic_res,
  effect_var,
  E_effect_var,
  cause_var_colname = "cause_var",
  include_var = "strongest_only",
  p_threshold = 0.05
)
}
\arguments{
\item{block}{Data.frame contains time series data.}

\item{uic_res}{Data.frame contains UIC results. Usually this is the output of \code{uic_across()}, but custom data.frame may be usable.}

\item{effect_var}{Character or Numeric. Column name or index of the effect variable.}

\item{E_effect_var}{Numeric. Optimal embedding dimension of the effect variable.}

\item{cause_var_colname}{Character. A column name for causal variables in \code{uic_res}.}

\item{include_var}{Character. \code{all_significant}, \code{strongest_only}, or \code{tp0_only}. If \code{all_significant}, all significantly influencing variables are used for the embedding. If \code{strongest_only}, tp with the strongest influence for each variable is used. If \code{tp0_only}, only variables with no time-delay are used.}

\item{p_threshold}{Numeric. Random seed.}
}
\value{
Embedded time series will be returned.
}
\description{
\code{make_block_mvd} Generate data.frame for MDR S-map
}
