% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_dir_handling.R
\name{outdir_create}
\alias{outdir_create}
\title{Create output directory using the currently-running file name}
\usage{
outdir_create(save_dir_name = TRUE, str_end = 3, suffix = "Out")
}
\arguments{
\item{save_dir_name}{Logical. If \code{TRUE}, output directory name is saved to an object.}

\item{str_end}{Integer. Trim \code{str_end} characters from the file name.}

\item{suffix}{Character. Suffix that will be appended to the file name.}
}
\value{
Output directory name.
}
\description{
\code{outdir_create} automatically detect the file name of the current srcipt and create output directory using the file name.
}
\examples{
# outdir_create()
}
