% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_based_rarefaction.R
\name{plot_rarefy}
\alias{plot_rarefy}
\title{Visualize rarefaction curve}
\usage{
plot_rarefy(ps_obj, plot_rarefied_point = TRUE, ran_seed = 1234)
}
\arguments{
\item{ps_obj}{Phyloseq object.}

\item{plot_rarefied_point}{Logical. Specify whether rarefied reads are plotted.}

\item{ran_seed}{Numeric. Random seed.}
}
\value{
ggplot object (\code{g_rare}). User may further edit the figure using \code{ggplot2}.
}
\description{
\code{plot_rarefy} visualizes a coverage-based rarefaction.
}
\examples{
# plot_rarefy(ps_obj)
}
