% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_based_rarefaction.R
\name{plot_rarefy}
\alias{plot_rarefy}
\title{Visualize rarefaction curve}
\usage{
plot_rarefy(
  ps_obj,
  plot_rarefied_point = TRUE,
  plot_slope = FALSE,
  linetype = 1
)
}
\arguments{
\item{ps_obj}{Phyloseq object.}

\item{plot_rarefied_point}{Logical. Specify whether rarefied reads are plotted.}

\item{plot_slope}{Logical. Specify whether tangent lines are plotted.}

\item{linetype}{Numeric or character. Specify the line type of the tangent lines.}
}
\value{
ggplot object (\code{g_rare}). User may further edit the figure using \code{ggplot2}.
}
\description{
\code{plot_rarefy} visualizes a coverage-based rarefaction.
}
\examples{
# plot_rarefy(ps_obj)
}
