% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer_hit.R
\name{PrimerHits}
\alias{PrimerHits}
\title{PrimerHits}
\usage{
PrimerHits(primer, fn)
}
\arguments{
\item{primer}{Primer sequence}

\item{fn}{File path to a FASTA/FASTQ file}
}
\value{
The number of primer hits.
}
\description{
\code{PrimerHits} search primer sequences in a FASTA/FASTQ file.
}
\examples{
# PrimerHits(primer, fn)
}
