% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_based_rarefaction.R
\name{rarefy_even_coverage}
\alias{rarefy_even_coverage}
\title{Perform coverage-based rarefaction for a phyloseq object}
\usage{
rarefy_even_coverage(
  ps_obj,
  coverage = 0.97,
  remove_not_rarefied = FALSE,
  include_iNEXT_results = FALSE,
  correct_singletons = FALSE,
  se = FALSE,
  nboot = 40,
  knots = 50,
  n_rarefy_iter = 1,
  rarefy_average_method = "round",
  sample_method = "phyloseq",
  ran_seed = 1234
)
}
\arguments{
\item{ps_obj}{Phyloseq object.}

\item{coverage}{Numeric. Coverage specified by a user (default = 0.97 = 97\%).}

\item{remove_not_rarefied}{Logical. If \code{TRUE}, samples of which coverage is lower than the specified coverage will be removed.}

\item{include_iNEXT_results}{Logical. If TURE, iNEXT results will be included as a element of the output. The first object is a rarefied phyloseq object, and the second object is an iNEXT result. Also, if \code{TRUE}, computation time will increase. If \code{FALSE}, it returns a rarefied phyloseq object only.}

\item{correct_singletons}{Logical. If TURE, singleton counts will be corrected according to the method of Chiu & Chao (2016). Corrected coverage and reads will be used for the rarefaction of the original data.}

\item{se}{Logical. Specify whether SE of the rarefaction curve is calculated.}

\item{nboot}{Numeric. Specify \code{nboot} of iNEXT function (valid only if \code{include_iNEXT_results = TRUE}).}

\item{knots}{Numeric. Specify \code{knots} of iNEXT function (valid only if \code{include_iNEXT_results = TRUE}).}

\item{n_rarefy_iter}{Numeric. The number of iterations of rarefactions (default = 1).}

\item{rarefy_average_method}{Character. If \code{n_rarefy_iter} >= 2, this argument determines how the multiple rarefactions are summarized. r\code{arefy_average_method = "round"} uses \code{round()}. \code{rarefy_average_method = "floor"} uses \code{floor()}. \code{rarefy_average_method = "ceiling"} uses \code{ceiling()}.}

\item{sample_method}{Character. Specify which function is used for rarefaction. \code{sample_method = "vegan"} uses \code{vegan::rrarefy()}, while \code{sample_method = "phyloseq"} uses \code{phyloseq:::rarefaction_subsample()}.}

\item{ran_seed}{Numeric. Random seed.}
}
\value{
Rarefied phyloseq object (\code{ps_rare}). If \code{include_iNEXT_results = TRUE}, \code{iNEXT} results are stored in the second element of the list.
}
\description{
\code{rarefy_even_coverage} performs coverage-based rarefaction for a phyloseq object using iNEXT package
}
\details{
\itemize{
\item{Hsieh et al. (2016) Methods in Ecology and Evolution https://doi.org/10.1111/2041-210X.12613}
}
}
\examples{
# rarefy_even_coverage(ps_obj, coverage = 0.97)
}
