% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr_smap.R
\name{s_map_mdr}
\alias{s_map_mdr}
\title{A function for MDR S-map}
\usage{
s_map_mdr(
  block_mvd,
  dist_w,
  lib = c(1, nrow(block_mvd)),
  pred = lib,
  tp = 1,
  theta = 8,
  weight_method = "sqrt",
  regularized = FALSE,
  lambda = 0,
  alpha = 0,
  glmnet_parallel = FALSE,
  save_smap_coefficients = FALSE,
  random_seed = 1234
)
}
\arguments{
\item{block_mvd}{Data.frame contains time series data. The first column should be the target column.}

\item{dist_w}{Matrix. Multiview disntance matrix. Usually this is the output of \code{compute_mvd()}.}

\item{lib}{Numeric vector. Library indices.}

\item{pred}{Numeric vector. Prediction indices.}

\item{tp}{Numeric. Forecasting time ahead.}

\item{theta}{Numeric. Weighing function for S-map.}

\item{weight_method}{Specify weighing method for \code{dist_w}. Default is \code{sqrt}, which is the same for Chang et al. (2021).}

\item{regularized}{Logical If \code{TRUE}, regularized S-map will be performed. If \code{FALSE}, the normal S-map will be performed. Please use \code{rEDM::s_map} function.}

\item{lambda}{Numeric. Specify the strength of penalty in the regularization.}

\item{alpha}{Numeric. \code{alpha = 0} is the ridge regression, \code{alpha = 1} is the lasso regression, and \verb{0 < alpha < 1} is an elastic net.}

\item{glmnet_parallel}{Logical. If TRUE, the computation will be parallel (currently, experimental).}

\item{save_smap_coefficients}{Logical. If \code{TRUE}, S-map coefficients will be saved.}

\item{random_seed}{Numeric. Random seed.}
}
\value{
A list that contains predictions, statistics, and S-map coefficients (if \code{save_smap_coefficients = TRUE})
}
\description{
\code{s_map_mdr} performs MDR S-map introduced in Chang et al. (2021) Ecology Letters.
}
\details{
\itemize{
\item{Chang, C.-W., Miki, T., Ushio, M., Ke, P.-J., Lu, H.-P., Shiah, F.-K. & et al. (2021) Reconstructing large interaction networks from empirical time series data. Ecology Letters, 24, 2763– 2774. https://doi.org/10.1111/ele.13897}
}
}
