% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_folder_handling.R
\name{save_session_info}
\alias{save_session_info}
\title{Save session information}
\usage{
save_session_info(output_dir, session_info_dir = "00_SessionInfo")
}
\arguments{
\item{output_dir}{xxx.}

\item{session_info_dir}{xxx.}
}
\description{
\code{save_session_info} xxx
}
\examples{
# output_dir <- outdir_create()
# save_session_info(output_dir)
}
