% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_dir_handling.R
\name{save_session_info}
\alias{save_session_info}
\title{Save session information using an output directory name}
\usage{
save_session_info(
  session_info_dir = "00_SessionInfo",
  create_session_info_dir = FALSE,
  str_end = 3,
  suffix = NULL
)
}
\arguments{
\item{session_info_dir}{Character. Name of the session information directory.}

\item{create_session_info_dir}{Logical. If \code{TRUE}, the function automatically creates an output directory for the session information}

\item{str_end}{Integer. Trim \code{str_end} characters from the file name.}

\item{suffix}{Character. Suffix that will be appended to the file name.}
}
\description{
\code{save_session_info} saves session information and append date information automatically.
}
\examples{
# output_dir <- outdir_create()
# save_session_info(output_dir)
}
