% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_folder_handling.R
\name{save_workspace}
\alias{save_workspace}
\title{Save workspace}
\usage{
save_workspace(output_dir, list = ls(all.names = TRUE))
}
\arguments{
\item{output_dir}{xxx.}

\item{list}{xxx.}
}
\description{
\code{save_workspace} xxx
}
\examples{
# output_dir <- outdir_create()
# save_workspace(output_dir)
}
