% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr_smap.R
\name{uic_across}
\alias{uic_across}
\title{Perform UIC across data.frame}
\usage{
uic_across(
  block,
  effect_var,
  uic_method = "optimal",
  E_range = 0:10,
  tp_range = -4:0,
  random_seed = 1234,
  silent = FALSE
)
}
\arguments{
\item{block}{Data.frame contains time series data.}

\item{effect_var}{Character or Numeric. Column name or index of the effect variable.}

\item{uic_method}{Character. \code{optimal} or \code{marginal}. For detail, see https://github.com/yutakaos/rUIC.}

\item{E_range}{Numeric. Embedding dimensions that will be tested.}

\item{tp_range}{Numeric. \code{tp} tested for UIC.}

\item{random_seed}{Numeric. Random seed.}

\item{silent}{Logical. if \code{TRUE}, progress message will not be shown.}
}
\value{
A data.frame that contains UIC results
}
\description{
\code{uic_across} Perform UIC for a target variable and multiple causal variables
}
\details{
\itemize{
\item{Osada & Ushio (2020) rUIC:Unified Information-theoretic Causality for R. https://doi.org/10.5281/zenodo.5163234}
}
}
